+++ src/components/CurrentUser/CurrentUser.tsx
@@ -1,7 +1,7 @@
 import React, { useEffect, useState } from "react";
 import { useSelector } from "react-redux";
 import { listSelector } from "../../store/store";
-import { Person } from "../../store/types";
+import { Person } from "../../interfaces";
 import { pullLocalStorage, pushLocalStorage } from "../../utils/localStorage";
 
 const localStorageKey = "currentUserIdMemory";
