+++ src/interfaces/Person.ts
@@ -1,0 +1,4 @@
+export interface Person {
+  readonly id: number;
+  readonly name: string;
+}
