+++ src/interfaces/Store.ts
@@ -1,0 +1,6 @@
+import { List } from "./List";
+
+export interface Store {
+  readonly list: List | null;
+  readonly currentListKey: List["key"] | null;
+}
