+++ src/interfaces/index.ts
@@ -1,0 +1,4 @@
+export * from "./List";
+export * from "./Person";
+export * from "./Item";
+export * from "./Store";
